:recycle: #24 refactor controller
improve path configuration
improve path mappings
