+++ src/main/java/pcc/puppet/enforcer/realm/organization/api/OrganizationController.java
@@ -16,7 +16,6 @@
 package pcc.puppet.enforcer.realm.organization.api;
 
 import static pcc.puppet.enforcer.realm.configuration.HttpHeaders.REQUESTER;
-import static pcc.puppet.enforcer.realm.organization.api.OrganizationController.BASE_PATH;
 
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.http.MediaType;
@@ -25,7 +24,6 @@ import io.micronaut.http.annotation.Controller;
 import io.micronaut.http.annotation.Get;
 import io.micronaut.http.annotation.Header;
 import io.micronaut.http.annotation.Post;
-import io.micronaut.http.annotation.QueryValue;
 import io.micronaut.security.annotation.Secured;
 import io.micronaut.security.rules.SecurityRule;
 import io.micronaut.tracing.annotation.NewSpan;
@@ -41,12 +39,10 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @Slf4j
-@Controller(BASE_PATH)
+@Controller("${micronaut.http.services.organization.path}")
 @Secured(SecurityRule.IS_ANONYMOUS)
 @RequiredArgsConstructor
 public class OrganizationController implements OrganizationOperations {
-
-  public static final String BASE_PATH = "/realm/organization";
   private final OrganizationService organizationService;
 
   @NewSpan
@@ -58,18 +54,18 @@ public class OrganizationController implements OrganizationOperations {
   }
 
   @NewSpan
-  @Get(produces = MediaType.APPLICATION_JSON)
+  @Get(uri = "/{organizationId}", produces = MediaType.APPLICATION_JSON)
   public Mono<OrganizationPresenter> findOrganization(
       @NonNull @SpanTag(REQUESTER) @Header(REQUESTER) String requester,
-      @NonNull @SpanTag @QueryValue("organizationId") String organizationId) {
+      @NonNull @SpanTag String organizationId) {
     return organizationService.findById(requester, organizationId);
   }
 
   @NewSpan
-  @Get(uri = "child", produces = MediaType.APPLICATION_JSON)
+  @Get(uri = "/{organizationId}/child", produces = MediaType.APPLICATION_JSON)
   public Flux<OrganizationPresenter> findChildOrganizations(
       @NonNull @SpanTag(REQUESTER) @Header(REQUESTER) String requester,
-      @NonNull @SpanTag @QueryValue("organizationId") String organizationId) {
+      @NonNull @SpanTag String organizationId) {
     return organizationService.findByParentId(requester, organizationId);
   }
 }
